address review and use savepoints
