+++ coordinatorlib/src/main/scala/com/socrata/datacoordinator/truth/universe/sql/PostgresUniverse.scala
@@ -13,10 +13,10 @@ import com.socrata.datacoordinator.truth.metadata._
 import com.socrata.datacoordinator.truth.sql.{PostgresDatabaseMutator, PostgresDatabaseReader, RepBasedSqlDatasetContext, SqlColumnRep}
 import com.socrata.datacoordinator.truth._
 import com.socrata.datacoordinator.truth.metadata.sql._
-import com.socrata.datacoordinator.secondary.{CollocationManifest, PlaybackToSecondary, SecondaryMoveJobs, SecondaryManifest, SecondaryReplicationMessages}
+import com.socrata.datacoordinator.secondary._
 import com.socrata.datacoordinator.truth.loader._
 import com.socrata.datacoordinator.truth.loader.sql._
-import com.socrata.datacoordinator.secondary.sql.{SqlCollocationManifest, SqlSecondaryMoveJobs, SqlSecondaryManifest, SqlSecondaryStoresConfig}
+import com.socrata.datacoordinator.secondary.sql._
 import com.socrata.datacoordinator.util._
 import com.socrata.datacoordinator.util.collection.ColumnIdMap
 import org.slf4j.LoggerFactory
@@ -157,7 +157,7 @@ class PostgresUniverse[ColumnType, ColumnValue](conn: Connection,
     new SqlSecondaryManifest(conn)
 
   lazy val collocationManifest: CollocationManifest =
-  new SqlCollocationManifest(conn)
+    new PostgresCollocationManifest(conn)
 
   lazy val secondaryMoveJobs: SecondaryMoveJobs =
     new SqlSecondaryMoveJobs(conn)
